📝 refactor Firebase configuration path and initialize Firestore in the lib directory
