+++ src/hooks/userFetchDocument.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react"
-import { db } from "../firebase/config"
+import { db } from "../lib/firebase/config"
 import { doc, getDoc } from 'firebase/firestore'
 
 export const userFetchDocument = (docCollection, id) => {
