+++ src/hooks/userUpdateDocument.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect, useReducer } from "react";
-import { db } from "../firebase/config";
+import { db } from "../lib/firebase/config";
 import { updateDoc, doc } from "firebase/firestore";
 
 const initialState = {
